Problem fixed, still needs tidy up
