+++ src/NoteQuestionParser.ts
@@ -36,8 +36,6 @@ export class NoteQuestionParser {
         onlyKeepQuestionsWithTopicPath: boolean,
     ): Promise<Question[]> {
         this.noteFile = noteFile;
-        const noteText: string = await noteFile.read();
-
         // For efficiency, we first get the tag list from the Obsidian cache
         // (this only gives the tag names, not the line numbers, but this is sufficient for this first step)
         const tagList: string[] = noteFile.getAllTagsFromCache();
@@ -47,10 +45,13 @@ export class NoteQuestionParser {
         testTimeLog("cql.A");
 
         if (this._hasTopicPaths) {
+            const noteText: string = await noteFile.read();
+
+            testTimeLog(`cql.B\t${noteText.length}`);
             // Now that we know there are relevant flashcard tags in the file, we can get the more detailed info
             // that includes the line numbers of each tag
             const tagCacheList: TagCache[] = noteFile.getAllTagsFromText();
-            testTimeLog("cql.B");
+            testTimeLog("cql.C1");
 
             // The following analysis can require fair computation.
             // There is no point doing it if there aren't any topic paths
@@ -74,9 +75,11 @@ export class NoteQuestionParser {
                 this.questionList = this.questionList.filter((q) => q.topicPathList);
             }
         } else {
+            testTimeLog("cql.B");
+            testTimeLog("cql.C2");
             this.questionList = [] as Question[];
         }
-        testTimeLog("cql.C");
+        testTimeLog("cql.D");
         return this.questionList;
     }
