+++ src/util/DateProvider.ts
@@ -2,32 +2,6 @@ import moment from "moment";
 import { Moment } from "moment";
 import { ALLOWED_DATE_FORMATS } from "src/constants";
 
-var testTimeInfo: [ string, number ][];
-
-export function testTimeStart(): void {
-    testTimeInfo = [ [ "Start", moment().valueOf() ] ];
-}
-
-export function testTimeLog(desc: string): void {
-    testTimeInfo.push([ desc, moment().valueOf() ]);
-}
-
-export function testTimeGetLapTime(): number {
-    return moment().valueOf() - testTimeInfo[0][1];
-}
-
-export function testTimeFormatLapInfo(): string {
-    let prevTime: number = testTimeInfo[0][1];
-    let result: string = "";
-    for (let i = 1; i < testTimeInfo.length; i++) {
-        const thisTime: number = testTimeInfo[i][1];
-        result += `\t${testTimeInfo[i][0]}\t${thisTime - prevTime}`;
-        prevTime = thisTime;
-    }
-    result += `\tLapTime\t${testTimeGetLapTime()}|`;
-    return result;
-}
-
 export interface IDateProvider {
     get today(): Moment;
 }
